RT-API-Testing: small refactoring
