+++ api/json-task/helper/json-helper.js
@@ -1,0 +1,10 @@
+const fs = require('fs');
+
+module.exports = {
+    serializeCompanies: function (companies) {
+        return JSON.parse(JSON.stringify(companies));
+    },
+    readExpectedJson: function (filePath) {
+        return JSON.parse(fs.readFileSync(filePath, 'utf-8'));
+    },
+};
